drivers/fs/ext2: extent dispatch + 64-byte group descs + incompat gate
Wire ext3/4 support end-to-end inside the driver: file-block
lookup dispatches to extent-tree walk when EXTENTS is set,
group descriptors decode 64-byte ext4 layout when 64BIT is set,
and unsupported incompat features now refuse the mount instead
of silently corrupting the read path.

volume.rs::map_block:
- Return type widened u32 → u64 (ext4 extents carry 48-bit phys).
- Dispatches on Superblock::uses_extents():
  - true: new map_block_extents path. Serialises Inode::block[15]
    back to 60 LE bytes (the same bytes ext4 packs as an extent
    root in the inode), runs lookup_in_node, recurses on
    DeeperLookupRequired by reading the child extent block.
    Depth budget = 8 (spec caps at 5; paranoid extra slack).
    Uninitialized extents surface as physical=0 → caller
    zero-fills, matching the sparse-hole shape.
  - false: legacy 12-direct + 3-indirect walk (unchanged shape;
    each return widened to `as u64`).

volume.rs mount path:
- Calls superblock.check_incompat_features() before the BGDT
  read. Encryption / inline-data / casefold / etc. → refused.
- Uses superblock.effective_desc_size() to size the BGDT read
  and to pick GroupDesc::parse_sized(buf, desc_size) — 32 bytes
  for ext2/3, 64 bytes for ext4 64BIT. Each descriptor's hi+lo
  pairs are now assembled into u64 inode_table / block_bitmap /
  inode_bitmap fields. Existing call sites (gd.inode_table for
  the inode-table base) work unchanged since they cast to u64.

inode.rs is unchanged — the i_block[15] array is what stores
the extent root for ext4, just interpreted differently in the
new map_block path.

node.rs: read_block now takes the u64 phys map_block returns
without further casts.

One new smoke under drivers/fs/ext2:
- Round-trips an extent-root through the [u32; 15] i_block
  packing/unpacking + lookup_in_node, proves the inode-to-extent-
  parser handoff is byte-stable.

What's deliberately deferred (was task #75's third item):
- Factory registration with root_mount. Blocked on a Sync→Async
  block-device adapter: Ext2Volume::mount needs a BlockDevice
  (async, impl Future, not dyn-compatible) but the FsFactory
  signature in narf-filesystem::root_mount takes
  Arc<dyn BlockDeviceSync> (sync). A bridge that wraps a sync
  block device as an async one + spawns the mount future is
  larger than the dispatch fix; lands separately.

  The dispatch + incompat gate are the load-bearing changes
  that turn ext2 driver → ext4 driver — factory registration
  is just plumbing once that bridge crate exists.

Suite: 2223 pass / 0 fail / 50 skip.

+++ drivers/fs/ext2/src/node.rs
@@ -146,7 +146,7 @@ impl<B: BlockDevice + 'static> Ext2Node<B> {
                 }
             } else {
                 let mut blockbuf = vec![0u8; bs as usize];
-                self.volume.read_block(physical as u64, &mut blockbuf).await?;
+                self.volume.read_block(physical, &mut blockbuf).await?;
                 dst[total..total + chunk]
                     .copy_from_slice(&blockbuf[in_block..in_block + chunk]);
             }
